refactor: move icm translate loader and translations type to own files
