+++ src/app/core/utils/translate/translations.type.ts
@@ -1,0 +1,1 @@
+export type Translations = Record<string, string | Record<string, string>>;
